move as much functionality into the base cloudsql model class as possible; move close of gcs_wrapper to after upload_run_file()
